Ajustes das Estruturação do Projeto e Config do BD
