+++ bd/configdb.js
@@ -1,0 +1,17 @@
+// const configDB = {
+//     user: 'postgres',
+//     host: 'localhost',
+//     database: 'dbESPM',
+//     password: 'admin',
+//     port: 5432,
+// };
+
+const configDB = {
+    user: 'agoyxjdg',
+    host: 'isilo.db.elephantsql.com',
+    database: 'agoyxjdg',
+    password: 'Msfo_S2mmh736CABKE_FEwSv6HVIwasr',
+    port: 5432,
+};
+
+module.exports = configDB;
