+++ bd/index.js
@@ -7,5 +7,9 @@ var pool = new Pool(configdb);
 module.exports = {
     query: (text,params,callback) => {
         return (pool.query(text,params,callback));
-    }
+    },
+    viaCPF: " SELECT * from CURSOS where 1 = $1::integer " +
+    " " +
+    " ",
+    viaCOD: ""
 }
