--- db/configdb.js
@@ -1,9 +1,0 @@
-const configDB = {
-    user: 'postgres',
-    host: 'localhost',
-    database: 'MeuDB',
-    password: 'admin',
-    port: 5432,
-};
-
-module.exports = configDB;
