+++ public/stylesheets/style.css
@@ -1,3 +1,6 @@
 body{
     background-color: beige;
 }
+table, th, td {
+  border: 1px solid black;
+}
