+++ views/search.pug
@@ -1,0 +1,57 @@
+html
+    head
+        title Base de Concursos
+        link(rel="stylesheet", href="../public/stylesheets/style.css")
+    body
+        a(href="./") PÁGINA INICIAL
+        p BASE DE CONCURSOS
+
+    form(action="/search", method="post")
+        label Termo:
+        input(type="text" name="termo" required)
+        br
+        label Opcao:
+        select(name="opcao" required)
+            option(value="1") CPF
+            option(value="2") Profissao
+        input(type="submit" value="Enviar")
+
+    if data.length > 0
+        h4 Resultados
+            table
+                thead
+                    tr
+                        td Categoria
+                        td Opcao
+                tbody
+                each val in data    
+                    tr
+                        td=val.nome
+                        td=val.categoria
+
+    //- if opcao = 1
+    //-     table
+    //-         thead
+    //-             tr
+    //-                 td Órgão
+    //-                 td Código
+    //-                 td Edital
+    //-         tbody
+    //-         each val in data    
+    //-             tr
+    //-                 td=val.orgao
+    //-                 td=val.codigo
+    //-                 td=val.edital
+    //- else
+    //-     table
+    //-         thead
+    //-             tr
+    //-                 td Nome
+    //-                 td Data Nascimento
+    //-                 td CPF
+    //-         tbody
+    //-         each val in data    
+    //-             tr
+    //-                 td=val.nome
+    //-                 td=val.datanascimento
+    //-                 td=val.cpf
